[#7] refact : function folder
