+++ src/functions/typeChecker.ts
@@ -1,0 +1,13 @@
+export const typeChecker = (type: string) => {
+  try {
+    let isValid = true;
+    if (type != 'ticker' && type != 'orderbook' && type != 'trade') {
+      isValid = false;
+    }
+    return isValid;
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+export default typeChecker;
